% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airports}
\alias{airports}
\title{Data about the airports in US}
\format{
## `airports`
A data frame with 1,458 rows and 8 columns:
\describe{
  \item{faa}{The unique ID of airport}
  \item{name}{Airport Name}
  \item{lat, lon}{The position of airports}
  ...
}
}
\source{
<https://github.com/tidyverse/nycflights13>
}
\usage{
airports
}
\description{
Data about the airports in US
}
\keyword{datasets}
