% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.r
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficients of linreg object}
\usage{
\method{coef}{linreg}(x)
}
\arguments{
\item{x}{linreg object}
}
\value{
float list
}
\description{
Returns all the coefficients of the linreg object
}
