% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef}
\alias{coef}
\title{Title}
\usage{
coef(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Additional arguments}
}
\value{
The result of the method applied to the class of the input object.
}
\description{
Title
}
