% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ridgereg.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{The coefficients from the Ridge regression  of class 'ridgereg'}
\usage{
\method{coef}{ridgereg}(l)
}
\description{
The coef() returns all estimates for the coefficients in your Ridge regression.
}

