% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linreg.R
\docType{data}
\name{formula}
\alias{formula}
\title{linreg}
\format{\preformatted{Class 'formula' length 3 iris$Sepal.Length ~ iris$Petal.Length
  ..- attr(*, ".Environment")=<environment: namespace:Lab4> 
}}
\usage{
formula
}
\arguments{
\item{a}{"formula" of the form y ~ x+... and a dataframe}
}
\value{
an object with class "linreg", i.e. that has attributes according to linear regression.
}
\description{
mimics the lm-function in R, i.e. makes linear regression.
}
\author{
Niclas Lovsjö & Maxime Bonneau
}
\references{
https://en.wikipedia.org/wiki/Linear_regression
}
\keyword{datasets}

