% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{iris}
\alias{iris}
\title{Edgar Anderson's Iris Data}
\format{\code{iris} is a data frame with 150 cases (rows) and 5 variables
  (columns) named \code{Sepal.Length}, \code{Sepal.Width}, \code{Petal.Length},
  \code{Petal.Width}, and \code{Species}.}
\source{
Fisher, R. A. (1936) The use of multiple measurements in
  taxonomic problems. Annals of Eugenics, 7, Part II, 179–188.

  The data were collected by Anderson, Edgar (1935). The irises of
  the Gaspe Peninsula, Bulletin of the American Iris Society, 59, 2–5.
}
\usage{
iris
}
\description{
This famous (Fisher's or Anderson's) iris data set gives the
measurements in centimeters of the variables sepal length and
width and petal length and width, respectively, for 50 flowers
from each of 3 species of iris. The species are Iris setosa,
versicolor, and virginica.
}
\keyword{datasets}
