% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-package.R
\docType{package}
\name{lab4-package}
\alias{lab4}
\alias{lab4-package}
\title{lab4: Lab4}
\description{
Assignment for lab 4 in Advanced R programming. This package is used to fit lienar models.
}
\keyword{internal}
