% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4.R
\docType{package}
\name{lab4}
\alias{lab4}
\alias{lab4-package}
\title{A package for creating linreg RC objects.}
\description{
This package creates a linreg object. The object contains the qr matrix, coefficients, fitted values, residuals, degrees of freedom, variance, t_values and p_values of the linear model for \code{x} and \code{y}. The object needs aid from ggplot2 to make the graphics inside linreg$plot().
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula object like x ~ y.}

\item{\code{data}}{A data frame which contains the formula variables.}

\item{\code{X}}{The matrix created from formula.}

\item{\code{qr}}{The QR matrix}

\item{\code{y}}{A vector with the independent variable values}

\item{\code{Q}}{The matrix Q}

\item{\code{R}}{The matrix R}

\item{\code{coefficients}}{A vector with the beta coefficients}

\item{\code{fitted_values}}{A vector with the predicted values of y.}

\item{\code{residuals}}{A vector with the residuals (y - fitted_values).}

\item{\code{df}}{The degrees of freedom.}

\item{\code{Bvar}}{A vector containing the variance of the coefficients.}

\item{\code{resstdev}}{A vector with the standard deviation of the residuals}

\item{\code{t_values}}{A vector with the t_values of the coefficients.}

\item{\code{p_values}}{A vector with the p_values of the coefficients.}
}}

\references{
More information of QR decomposition \href{https://en.wikipedia.org/wiki/QR_decomposition}{here}.

More information on Linear regression \href{https://en.wikipedia.org/wiki/Linear_regression}{here}.
}
\author{
\strong{Maintainer}: Jose Mendez \email{josme478@student.liu.se}

Authors:
\itemize{
  \item Sara Salahshour \email{sarsa293@student.liu.se}
  \item Atieh Khaleghi \email{atikh281@student.liu.se}
}

}
