\name{lab4ab-package}
\alias{lab4ab-package}
\alias{lab4ab}
\docType{package}
\title{
\packageTitle{lab4ab}
}
\description{
\packageDescription{lab4ab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab4ab}
\packageIndices{lab4ab}

}
\author{
\packageAuthor{lab4ab}

Maintainer: \packageMaintainer{lab4ab}
}
\references{

}
\keyword{ package }
\seealso{

}
\examples{

}
