% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_disc.R
\docType{package}
\name{Lab4Pack}
\alias{Lab4Pack}
\title{Linreg and theme for Linköping University}
\description{
Computes linear regression by matrix multiplication of least squares and a theme for Linköping University.
}
\author{
Leonard Persson Norblad \email{leope744@student.liu.se}

Jacob Sibanda \email{jacsi259@student.liu.se}
}
