% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\docType{class}
\name{LinearRegression-class}
\alias{.linreg}
\alias{LinearRegression-class}
\title{title here}
\source{
\url{https://stat.ethz.ch/R-manual/R-devel/library/methods/html/refClass.html}
}
\arguments{
\item{param}{description}
}
\value{

}
\description{
description here
}
\examples{
examples here

}

