% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_model.R
\docType{class}
\name{LinearRegressionModel}
\alias{LinearRegressionModel}
\alias{linregmod}
\title{A Reference Class to represent linear regression models.}
\source{
\url{https://en.wikipedia.org/wiki/Linear_regression}
}
\description{
Stores information about a linear regression model generated by \code{linreg}.
}
\section{Fields}{

\describe{
\item{\code{call}}{Object of class \code{"character"}. A string representation of the calling code.}

\item{\code{coefficients}}{Object of class \code{"list"}. A list containing information about the estimated coefficients.
\describe{
\item{\code{val}:}{Object of class \code{"numeric"}. The estimated parameters.}
\item{\code{se}:}{Object of class \code{"numeric"}. The standard errors of the parameters.}
\item{\code{var}:}{Object of class \code{"numeric"}. The variance of the parameters.}
\item{\code{tval}:}{Object of class \code{"numeric"}. The t-values of the the parameters.}
\item{\code{pval}:}{Object of class \code{"numeric"}. The p-values of the parameters.}
}}

\item{\code{residuals}}{Ojbect of class \code{"list"}. A list containing information about the residuals.
\describe{
\item{\code{val}:}{Object of class \code{"numeric"}. The residuals.}
\item{\code{var}:}{Object of class \code{"numeric"}. The variance of the residuals.}
}}

\item{\code{fitted.values}}{Object of class \code{"numeric"}. The fitted values of y using the model's parameters.}

\item{\code{df}}{Object of class \code{"numeric"}. The number of degrees of freedom in the model.}
}}
\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the estimated parameters.}

\item{\code{plot()}}{Plots Residuals vs Fitted and Scale-Location.}

\item{\code{pred()}}{Returns the fitted values.}

\item{\code{print()}}{Prints the model.}

\item{\code{resid()}}{Returns the residuals.}
}}

