% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{linear regression using QR}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A model formula}

\item{data}{A data frame containing the model variables.}
}
\value{
An object of class linreg containing the regression results.
}
\description{
linear regression using QR
}
