% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg_f}
\alias{linreg_f}
\title{Linear Regression Function}
\usage{
linreg_f(formula, data)
}
\arguments{
\item{formula}{formula (y ~ x)}

\item{data}{dataframe}
}
\value{
An object of RC class linreg
}
\description{
This function finds the shortest distance from a source
             node to all the other nodes in a graph
}
\examples{
linreg_f(formula = Petal.Length ~ Species, data = iris)
}
