% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_ols.R
\name{linreg_ols}
\alias{linreg_ols}
\title{Linear regression with ols method}
\usage{
linreg_ols(formula, data)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted}

\item{data}{A data frame containing the variables in the model}
}
\value{
statistics
}
\description{
Linear regression with ols method
}
