% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_qr.R
\name{linreg_qr}
\alias{linreg_qr}
\title{Fit a linear regression model (QR decomposition)}
\usage{
linreg_qr(formula, data)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data frame containing the variables in the model.}
}
\value{
An object of class `linreg` with QR-based estimates.
}
\description{
This function fits a multiple linear regression model using
a QR decomposition rather than the normal equations.
}
\examples{
data(iris)
fit <- linreg_qr(Petal.Length ~ Sepal.Length + Sepal.Width, data = iris)
summary(fit)
}
