% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linregmod.R
\docType{class}
\name{linregmod}
\alias{linregmod}
\title{A Reference Class to represent linear regression models.}
\source{
\url{https://stat.ethz.ch/R-manual/R-devel/library/methods/html/refClass.html}
}
\description{
Stores information about a linear regression model generated by \code{linreg}.
}
\section{Fields}{

\describe{
\item{\code{call}}{Object of class \code{"character"}. A string representation of the calling code.}

\item{\code{coefficients}}{Object of class \code{"list"}. A list containing information about the estimated coefficients.
\describe{
\item{\code{val}:}{Object of class \code{"numeric"}, the estimated parameters} 
}}
}}
\examples{
examples here

}

