% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_graph.R
\name{LiU_theme}
\alias{LiU_theme}
\title{Graphics for Linköping University}
\usage{
LiU_theme()
}
\description{
A ggplot2 theme for LiU. Recommend to render plot by cairo_pdf device in 4 x 6 dimensions.
}
\examples{
# data(iris)
#  ggplot(iris, aes(y=Sepal.Length, x=Petal.Length)) + 
#  geom_point(size=1) + 
#  labs(title="Scatterplot of Sepal Length and Petal Length",
#       y="Sepal Length",
#       x="Penatal Length") +
#  LiU_theme()


}
