% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.linreg}
\alias{plot.linreg}
\title{plot function}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class "linreg".}

\item{...}{Not used.}
}
\value{
The two plots, invisibly, as a list.
}
\description{
plot function
}
