% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pred.linreg}
\alias{pred.linreg}
\title{pred() should return the predicted values yˆ}
\usage{
pred.linreg(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to methods.}
}
\description{
pred() should return the predicted values yˆ
}
