% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pred}
\alias{pred}
\title{Predicted values
Generic function that returns the predicted values.}
\usage{
pred(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to methods.}
}
\value{
The predicted values.
}
\description{
Predicted values
Generic function that returns the predicted values.
}
