% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.ridgereg.R
\name{pred.ridgereg}
\alias{pred.ridgereg}
\title{The fitted values from the Ridge regression 'ridgereg'}
\usage{
\method{pred}{ridgereg}(l, newdata = NA)
}
\description{
The pred()-statement returns a vector including all fitted values from the Ridge regression.
}

