% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeReg.R
\name{predict}
\alias{predict}
\alias{predict.ridgereg}
\title{Predicted values for a model}
\usage{
predict(object, ...)

\method{predict}{ridgereg}(object, newdata, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Arguments passed to methods.}

\item{newdata}{A data frame for which to predict values.}
}
\description{
Returns the predicted values from a model.
}
