% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ridgereg.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{The coefficients from the Ridge regression 'ridgereg'.}
\usage{
\method{print}{ridgereg}(l)
}
\description{
The print() returns the names and the regression coefficients.
}

