% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{resid.linreg}
\alias{resid.linreg}
\title{residuals}
\usage{
\method{resid}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg object.}

\item{...}{extra arguments.}
}
\value{
vector of residuals.
}
\description{
residuals
}
