% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeReg.R
\name{ridgereg}
\alias{ridgereg}
\title{Fit a ridge regression model}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data frame containing the variables in the model.}

\item{lambda}{A numeric value for the regularization parameter.}
}
\value{
An object of class `ridgereg` containing:
\item{coefficients}{Estimated regression coefficients (un-scaled).}
\item{fitted.values}{Fitted (predicted) values.}
\item{residuals}{Residuals.}
\item{formula}{The model formula.}
\item{lambda}{The regularization parameter used.}
}
\description{
This function fits a multiple linear regression model using
ridge regularization and QR decomposition.
}
\examples{
data(iris)
fit_ridge <- ridgereg(Petal.Length ~ Sepal.Length + Sepal.Width, data = iris, lambda = 0.5)
print(fit_ridge)
}
