% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_colour_liu}
\alias{scale_colour_liu}
\title{LiU colours}
\usage{
scale_colour_liu(...)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
LiU colours
}
