% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.r
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary of linreg object}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{linreg object}

\item{...}{other options}
}
\description{
Prints on screen a summary of the different parameters of the linreg object.
This includes the estimate, the standard error, t value and p value for each
variable, as well as the estimated variance and the degrees of freedom.
}
