% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme}
\alias{theme}
\title{Linköping University theme for ggplot2}
\usage{
theme(base_size = 10)
}
\arguments{
\item{base_size}{Numeric. Base font size used for text elements
in the plot.}
}
\value{
A list of ggplot2 components that can be added to a
ggplot object using the `+` operator.
}
\description{
Creates a custom ggplot2 theme based on the visual identity of
Linköping University.
}
\references{
\url{https://www.ida.liu.se/~732A94/wwwfiles/filer/LiU_grafisk_manual.pdf}
}
