% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_liu}
\alias{theme_liu}
\title{Theme for LiU}
\usage{
theme_liu(base_size = 12, heading_family = "Calibri", body_family = "Georgia")
}
\arguments{
\item{base_size}{Base font size.}

\item{heading_family}{Font for title, subtitle, axis and legend titles.}

\item{body_family}{Font for all other text.}
}
\description{
From the LiU graphical manual: LiU colours,
Korolev LiU for headings and Miller for body text (Calibri and
Georgia).
}
