% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airportDelay.R
\name{visualize_airport_delay}
\alias{visualize_airport_delay}
\title{visualize_airport_delay}
\usage{
visualize_airport_delay()
}
\description{
Visualize airport delay use ridge regression to predict the delays in each airport.
}
