% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean_flight_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize Mean Flight Delays by Airport}
\usage{
visualize_airport_delays()
}
\value{
A \pkg{ggplot2} plot showing airports positioned by their geographic
coordinates.
}
\description{
Shows the mean flight delays for different destination
airports in the United States using data from the \pkg{nycflights13} package.
It combines information from the \code{flights} and \code{airports} datasets,
calculates the mean arrival delay for each destination airport, and plots
these delays by location.
}
\examples{
\dontrun{
visualize_airport_delays()
}

}
