% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Extract regression coefficients}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Additional arguments.}
}
\value{
A named numeric vector containing the estimated regression
coefficients.
}
\description{
Extracts the estimated regression coefficients from a \code{linreg} object.
}
