% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.linreg}
\alias{coef.linreg}
\title{coef() should return the coefficients as a named vector}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class "linreg".}

\item{...}{Not used.}
}
\value{
A named vector of regression coefficients.
}
\description{
coef() should return the coefficients as a named vector
}
