% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{The coefficients from the linear regression 'linreg'}
\usage{
\method{coef}{linreg}(l)
}
\description{
The coef() returns all estimates for the coefficients in your linear regression.
}

