% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.r
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficients of linreg object}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{linreg object}

\item{...}{other options}
}
\value{
float list
}
\description{
Returns all the coefficients of the linreg object
}
