% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{coefficients}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg object.}

\item{...}{extra arguments.}
}
\value{
vector of regression coefficients.
}
\description{
coefficients
}
