% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient.R
\name{coef.linreg}
\alias{coef.linreg}
\title{coef S3 method}
\usage{
\method{coef}{linreg}(object)
}
\arguments{
\item{linreg_obj}{An object of class linreg}
}
\value{
coefficients of linear regression
}
\description{
S3 method for objects of class linreg providing the coefficients of the linear regression
}
