% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{coefficients of the formula.}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Additional arguments}
}
\value{
object$Regressions_coefficients
}
\description{
coefficients of the formula.
}
