% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ridgereg.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{Coefficients of the formula}
\usage{
\method{coef}{ridgereg}(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Additional arguments}
}
\value{
object$coefficients
}
\description{
Coefficients of the formula
}
