% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris.R
\docType{data}
\name{iris}
\alias{iris}
\title{Edgar Anderson's Iris Data}
\format{
iris is a data frame with 150 cases (rows) and 5 variables (columns) named Sepal.Length, Sepal.Width, Petal.Length, Petal.Width, and Species. Iris3 gives the same data arranged as a 3-dimensional array of size 50 by 4 by 3, as represented by S-PLUS. The first dimension gives the case number within the species subsample, the second the measurements with names Sepal L., Sepal W., Petal L., and Petal W., and the third the species.
}
\source{
\url{https://www.rdocumentation.org/packages/datasets/versions/3.6.2/topics/iris}
}
\usage{
iris
}
\description{
This famous (Fisher's or Anderson's) iris data set gives the measurements in centimeters of the variables sepal length and width and petal length and width, respectively, for 50 flowers from each of 3 species of iris. The species are Iris setosa, versicolor, and virginica.
}
\details{
iris

column names: Sepal.Length, Sepal.Width, Petal.Length, Petal.Width, Species
}
\keyword{datasets}
