\name{lab4-package}
\alias{lab4-package}
\alias{lab4}
\docType{package}
\title{
\packageTitle{lab4}
}
\description{
\packageDescription{lab4}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{lab4}

Maintainer: \packageMaintainer{lab4}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = lab4)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{lab4}}
%\section{Documentation Index}{\packageIndices{lab4}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
