\name{Lab4-package}
\alias{Lab4-package}
\alias{Lab4}
\docType{package}
\title{
\packageTitle{Lab4}
}
\description{
\packageDescription{Lab4}
}

\author{
\packageAuthor{Lab4}

Maintainer: \packageMaintainer{Lab4}
}
\references{
https://en.wikipedia.org/wiki/Linear_regression
\keyword{ package }
\seealso{

}
\examples{
> linreg(Sepal.Width~Sepal.Length,iris)->x
> x
Call: 
linreg(Sepal.Width ~ Sepal.Length)

Coefficients: 
(Intercept) Sepal.Length 
3.418947 -0.0618848
}
