% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4.R
\docType{package}
\name{lab4}
\alias{lab4}
\alias{lab4-package}
\title{The Linear regression model package}
\description{
A package created as a part of Lab 4 in the course 732A94 @ LiU. This package implements some genereic functions for the class 
'linreg'. As well as a function that calculates the ordinary least squares-method 
and produces a list of the class 'linreg'.
}

