% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4-package.R
\docType{package}
\name{lab4}
\alias{lab4}
\title{A package for linear regression models}
\description{
This package offers implementation of a linear regression model and provides a dataset for the vignette.
}
\details{
lab4
}
