% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Lab4}
\alias{Lab4}
\alias{Lab4-package}
\title{Lab 4}
\description{
Implementation of Linear Regression and a Reference Class to store linear regression models.
}
\author{
Rasmus Holm

Zhendong Wang

Karolina Ziomek
}

