% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lab4.R
\docType{package}
\name{Lab4}
\alias{Lab4}
\alias{Lab4-package}
\title{Lab4 package contains a RC generator for linear regression analysis}
\description{
The Lab4 package consists of a single RC generator.
Package imports packages \code{ggplot2}, \code{grid} and \code{gridExtra}.
}
\section{Lab4 functions}{

The Lab4 function \code{linreg} creates an
\code{linreg} class object (actually an environment in R) which does linear regression analysis
upon input formula and data.frame. All the calculations are done and stored in the fields of the
object upon initialization. The methods just returns the values.
}

