% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{LinReg-class}
\alias{LinReg-class}
\alias{LinReg}
\title{Linear Regression}
\usage{
linreg(Petal.Length~Species, data = iris)
}
\arguments{
\item{formula}{The formula}

\item{data}{The data, iris}
}
\value{
Linear regression for Iris
}
\description{
Run Linear Regression
}

