% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Function for a multiple regression model}
\arguments{
\item{formula}{an object of class 'formula': a symbolic description of the model to be fitted.}

\item{data}{an 'data.frame' containing the variables in the model.}
}
\value{
Returns an object of class 'linreg'.
}
\description{
Package to handle linear regression models
}

\examples{
data("iris")
library(Lab4)
linear_model <- linreg(Petal.Length~Sepal.Width+Sepal.Length, data=iris)
}
