% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear regression using ordinary linear algebra}
\arguments{
\item{formula}{a formula object.}

\item{data}{a data frame.}
}
\description{
linreg is used t o fit linear models and uses the ordinary
  linear algebra to  calculate the regression
}

\examples{
# data(iris)
# l <- linreg$new(formula=Petal.Length~Sepal.Width+Sepal.Length, data=iris)
# l$summary()
 

}
