% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A RC class for multiple linear regression}
\description{
A RC class for multiple linear regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{The formula object containing depedent and independent variables}

\item{\code{data}}{A data frame object to apply the multiple linear regression to}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Returns linear regression coefficients}

\item{\code{pred()}}{Returns fitted values}

\item{\code{print()}}{Prints the input and the coefficients in a user-friedly way}

\item{\code{resid()}}{Returns residual vector}

\item{\code{summary()}}{Returns a summary of the linear regression}
}}

