% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{linreg}
\arguments{
\item{formula}{A list representing linear regression formula.}

\item{data}{The imput data for the model.}
}
\value{
A object which can return coefficients, residuals, degree of freedom and so on.
}
\description{
This function implements different linear regression algorithms.
}

\examples{
linreg(1, 2)

}
\references{
Wikipedia: \href{"https://en.wikipedia.org/wiki/Linear_regression"}{Linear Regression}
}
