% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{Linreg-class}
\alias{Linreg-class}
\alias{Linreg}
\title{linreg}
\arguments{
\item{formula}{A list representing linear regression formula.}

\item{data}{The imput data for the model.}

\item{which}{"which" determines which plot would be demostrated, when which is 1, "Residuals vs Fitted" demostrated
and when which is 3, "Scale-Location" demostrated.}
}
\value{
A object which can return coefficients, residuals, degree of freedom and so on.

The residual matrix of regression

The predicted value of y

The residual matrix of regression
}
\description{
This class implements different linear regression algorithms.

Function print() implements printing the coefficients matrix of regression.

Function plot() implements plots named "Residuals vs Fitted" and "Scale-Location".

Function resid() returns the residuals

Function pred() returns the predicted value of y

Function coef() returns the coefficients matrix

Function summary() prints a similar printout as printed for lm objects, but you only need to 
present the coeffcients with their standard error, t-value and p-value as well as the estimate of residual 
variance and the degrees of freedom in the model.
}

\examples{
data <- data.frame(y = c(1,2,3,4,5,6), x1 = c(7,6,5,4,3,2), x2 = c(9,8,7,6,5,4))
object<- Linreg$new(formula = y ~ x1 + x2, data = data)
object$print()
object$plot(which = 1)
object$plot(which = 3)
object$resid()
object$pred()
object$coef()
object$summary()
}
\references{
Wikipedia: \href{"https://en.wikipedia.org/wiki/Linear_regression"}{Linear Regression}
}
