% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{linreg}
\arguments{
\item{formula}{A list representing linear regression formula.}

\item{data}{The imput data for the model.}

\item{method}{This parameter determines the calculation method.}

\item{model}{This parameter determines which model you choose.}

\item{which}{"which" determines which plot would be demostrated, when which is 1, "Residuals vs Fitted" demostrated
and when which is 3, "Scale-Location" demostrated.}

\item{newData}{Calculate output based on new data}

\item{lambda}{Hyper parameter}
}
\value{
A object which can return coefficients, residuals, degree of freedom and so on.

The residual matrix of regression

The predicted value of y

The residual matrix of regression
}
\description{
This class implements different linear regression algorithms.

Function print() implements printing the coefficients matrix of regression.

Function plot() implements plots named "Residuals vs Fitted" and "Scale-Location".

Function resid() returns the residuals

Function pred() returns the predicted value of y

Function coef() returns the coefficients matrix

Function summary() prints a similar printout as printed for lm objects, 
presenting the coeffcients with their standard error, t-value and p-value as well as the estimate of residual 
variance and the degrees of freedom in the model.

This function is implemented for initializing the ridge regression model
}

