% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{LAB 4}
\value{
print() should print out the coeﬃcients and coeﬃcient names,similar as done by the lm class,
        plot() should plot the following two plots using ggplot2,
        resid() should return the vector of residuals e_hat,
        pred() should return the predicted values y_hat,
        coef() should return the coeﬃcients as a named vector,
        summary() should return a similar printout as printed for lm objects.
}
\description{
In this lab we will will create a package to handle linear regression models.
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{beta_hat}}{matrix.}

\item{\code{y_hat}}{matrix.}

\item{\code{e_hat}}{matrix.}

\item{\code{df}}{numeric.}

\item{\code{var_hat}}{numeric.}

\item{\code{p_value}}{matrix.}

\item{\code{t_value}}{matrix.}

\item{\code{var_beta_hat}}{matrix.}

\item{\code{data_name}}{character.}

\item{\code{res_val}}{matrix.}
}}


\examples{
linreg_mod <- linreg$new(Petal.Length~Sepal.Width+Sepal.Length,data=iris)
 linreg_mod$print()
 linreg_mod$pred()
 linreg_mod$resid()
 linreg_mod$coef()
 linreg_mod$summary()
 linreg_mod$plot()
          

}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Linear Regression Model}
}
