% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{lab4}
\description{
In this lab we will will create a package to handle linear regression models.
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{beta_hat}}{matrix.}

\item{\code{y_hat}}{matrix.}

\item{\code{e_hat}}{matrix.}

\item{\code{df}}{numeric.}

\item{\code{var_hat}}{numeric.}

\item{\code{p_value}}{matrix.}

\item{\code{t_value}}{matrix.}

\item{\code{var_beta_hat}}{matrix.}

\item{\code{data_name}}{character.}

\item{\code{res_val}}{matrix.}
}}


