% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg2.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{LAB 4}
\value{
an object with class linreg.
}
\description{
LAB 4
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{beta_hat}}{matrix.}

\item{\code{y_hat}}{matrix.}

\item{\code{e_hat}}{matrix.}

\item{\code{df}}{numeric.}

\item{\code{var_hat}}{numeric.}

\item{\code{p_value}}{matrix.}

\item{\code{t_value}}{matrix.}

\item{\code{var_beta_hat}}{matrix.}

\item{\code{data_name}}{character.}
}}


\examples{
linreg(Petal.Length~Species, data=iris)
}
