% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinReg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear regression by least squares}
\arguments{
\item{Data}{A data.frame}

\item{formula}{A formula}
}
\value{
\code{print()} Returns the function call and estimated coefficients

\code{plot()} Plots two ggplots. Residuals vs fitted and scale-location

\code{resid()} Returns a vector of residuals

\code{pred()} Returns a vector of predicted values

\code{coef()} Returns a vector of estimated coefficients

\code{summary()} Returns a short summary of the model estimation
}
\description{
Computes linear regression by matrix multiplication of least squares.
}

\examples{
example <- linreg(data=iris, formula=Petal.Length~Species)
example$summary()
example$print()

}
\references{
Edlund, O. (2011). Some Notes on Least Squares, OR-factorization, SVD and Fitting. \url{http://staff.www.ltu.se/~jove/courses/c0002m/least_squares.pdf}
}
