% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A reference class to perform linear regression and return parameters and plots.}
\description{
A reference class to perform linear regression and return parameters and plots.
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula: Object dependent and indepedent variables.}

\item{\code{data}}{data.frame: All the data.}

\item{\code{dname}}{vector: The names of the variables.}

\item{\code{X}}{matrix: Independent variables.}

\item{\code{y}}{matrix: Dependent variable.}

\item{\code{b_hat}}{matrix: Estimation of the regression coefficients.}

\item{\code{y_hat}}{matrix: Estimation of the y values.}

\item{\code{e_hat}}{matrix: Estimation of the error variable i.e. the residuals.}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Return the coefficient}

\item{\code{initialize(formula, data)}}{Initialize the values}

\item{\code{plot()}}{Plot the linear regression}

\item{\code{pred()}}{Return the predicted values}

\item{\code{print()}}{Print the coeff that needed}

\item{\code{resid()}}{Return the vector of residuals}

\item{\code{summary()}}{Summary of the linear regression}
}}

