% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A Reference Class to represent a linear model.}
\value{
An object of class " linreg "
}
\description{
A Reference Class to represent a linear model.
}
\section{Fields}{

\describe{
\item{\code{formula}}{Presents description of model to be fitted.}

\item{\code{data}}{Shows observations including dependent (Y) and independent variables (X).}

\item{\code{coefficients}}{The estimates for the model parameters to describe the relationship between X and Y.}

\item{\code{fittedvalues}}{The predicted value of Y.}

\item{\code{residuals}}{The difference between observed value of the Y and the value of Y variable predicted from the regression line.}

\item{\code{df}}{Difference between number of observations and parameters.}

\item{\code{residualvariance}}{The variance of the values that are calculated by finding the distance between regression line and the actual points.}

\item{\code{var_regcoef_matrix}}{The variance of the parameters that are calculated as coefficients.}

\item{\code{t_values}}{A matrix with the t-values of coefficient.}

\item{\code{p_value}}{The probability of finding the given t-statistic if the null hypothesis of no relationship between X and Y were true.}
}}


