% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Advanced Programming in R - Lab 4
Linear Regression using ordinary least squares}
\description{
Advanced Programming in R - Lab 4
Linear Regression using ordinary least squares
}

