% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A class for linear regression with many methods for corresponding values calculations}
\description{
A class for linear regression with many methods for corresponding values calculations
}
\section{Fields}{

\describe{
\item{\code{y}}{vector The orginial labels of observations in data}

\item{\code{X}}{matrix The features of all observations in data}

\item{\code{beta}}{vector The coefficients in the sample}

\item{\code{y_hat}}{matrix The estimated labels}

\item{\code{e_hat}}{matrix The standard errors between orginial and estimated labels}

\item{\code{df}}{numeric The degree of freedom of data}

\item{\code{var_resid}}{numeric The residual variance}

\item{\code{var_hat}}{matrix The variance of the regression coefficients}

\item{\code{t_val}}{vector The t-values for each coefficient}

\item{\code{p_val}}{vector The p-values for each coefficient}

\item{\code{formula}}{character The formula used in the model}

\item{\code{data}}{data.frame The sample data}
}}


