% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A Reference Class to represent a linear model.}
\description{
A Reference Class to represent a linear model.
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula object like x ~ y.}

\item{\code{data}}{A data frame which contains the formula variables.}

\item{\code{X}}{The matrix created from formula.}

\item{\code{qr}}{The QR matrix}

\item{\code{y}}{A vector with the independent variable values}

\item{\code{Q}}{The matrix Q}

\item{\code{R}}{The matrix R}

\item{\code{coefficients}}{A vector with the beta coefficients}

\item{\code{fitted_values}}{A vector with the predicted values of y.}

\item{\code{residuals}}{A vector with the residuals (y - fitted_values).}

\item{\code{df}}{The degrees of freedom.}

\item{\code{Bvar}}{A vector containing the variance of the coefficients.}

\item{\code{resstdev}}{A vector with the standard deviation of the residuals}

\item{\code{t_values}}{A vector with the t_values of the coefficients.}

\item{\code{p_values}}{A vector with the p_values of the coefficients.}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Prints the coefficients}

\item{\code{initialize(formula, data)}}{Creates a linear model based on formula and data}

\item{\code{plot()}}{Makes two plots}

\item{\code{pred()}}{Prints the predicted values}

\item{\code{print()}}{Prints the coefficients}

\item{\code{resid()}}{Prints the residuals}

\item{\code{summary()}}{Prints a summary table}
}}

