% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab4.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear model using QR decompoition}
\arguments{
\item{formula}{A formula with the shape x ~ y, where x is the dependent variable and y the independent variable.}

\item{data}{A dataset containing the variables of formula.}
}
\value{
A linreg Reference Class object containing the following statistics:

\code{coefficients} a named vector of coefficients,

\code{residuals} the residuals, that is response minus predicted values

\code{predicted_values} the fitted mean values

\code{variance} The variance

\code{t_values} t_value,

\code{p_value} p_value
}
\description{
This function creates a linreg object. The object contains the qr matrix, coefficients, fitted values, residuals, degrees of freedom, variance, t_values and p_values of the linear model for \code{x} and \code{y}.
}

\references{
More information of QR decomposition \href{https://en.wikipedia.org/wiki/QR_decomposition}{here}.

More information on Linear regression \href{https://en.wikipedia.org/wiki/Linear_regression}{here}.
}
