% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A RC class for linear regression models}
\description{
A RC class for linear regression models
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula}

\item{\code{data}}{A dataset which we apply the formula (data.frame)}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the regression coefficients}

\item{\code{plot()}}{Provides a plot for 'Scale-Location' and 'Residuals vs. Fitted',
returns a list of the two plots, accessed $p1 and $p2}

\item{\code{pred()}}{Returns fitted values}

\item{\code{print()}}{Prints the call of the class and regression coefficients in named vector}

\item{\code{resid()}}{Returns residual values}

\item{\code{summary()}}{Provides a summary of Estimate, standard error, t values and
p-values for linear regression coefficients}
}}

