% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg}
\alias{linreg-class}
\title{A reference class generator for linear regression}
\description{
This function creates a class \code{linreg} object which does linear
regression upon a specified formula and data.frame. Linear regression is
performed by ordinary least squares calculations. Plots are handled by
packages \code{ggplot2}, \code{grid} and \code{gridExtra}.
}
\details{
This RC class generator can create a \code{linreg} object from any formula as
long as the response are not factors. The ordinary least squares calculations
are all done using matrices, and the residual plots fit using \code{glm.fit}
by default, although it switches to \code{loess} for groups <1000 (see
\link[ggplot2]{stat_smooth}). The printouts are simple \code{paste} and
\code{writeLines} calls for the most part, and are not objects. Every method
returns something relevant.
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula taken as input, upon which linear regression is
  performed.}

\item{\code{data}}{A data.frame which has names to which the names in the formula
  refer.}

\item{\code{dataname}}{A character string containing the name of the data.}

\item{\code{coefficients}}{A single column matrix consisting of the calculated
  regression coefficients of the linear model.}

\item{\code{residual}}{A single column matrix containing the residuals of the
  calculated linear model.}

\item{\code{predicted}}{A single column matrix containing the linearly predicted
  response.}

\item{\code{df}}{An integer containing the degrees of freedom of the linear
  regression undertaken.}

\item{\code{residualvar}}{A numeric scalar indicating the residual variance of the
  linear model.}

\item{\code{varregcoefficients}}{A matrix containing the covariance matrix of the
  regression coefficients.}

\item{\code{tvalues}}{A single column matrix consisting of the t-statistics for the
  calculated regression coefficients.}

\item{\code{pvalues}}{A single column matrix consisting of the two-sided p-values
  associated with the t-statistics for the calculated regression
  coefficients. Is zero for large enough t-statistic.}
}}
\section{Methods}{

\describe{
\item{\code{coef()}}{Returns a named vector of the regression coefficients in the linear model.}

\item{\code{initialize(formula, data)}}{From inputs formula and data, generates all the other fields upon object initialization.}

\item{\code{plot()}}{plots two plots of residuals vs. fitted values, in a grid.}

\item{\code{pred()}}{Returns the predicted values of the linear model.}

\item{\code{print()}}{Gives a printout of the call as well as the calculated regression coefficients.}

\item{\code{resid()}}{Returns a vector of the residual values of the linear model.}

\item{\code{summary()}}{Prints out a summary of the linear regression performed.}
}}
\examples{
data(faithful)
linearobject <- linreg(formula = eruptions ~ waiting, data = faithful)
linearobject$summary()
}
\references{
\url{http://en.wikipedia.org/wiki/Linear_regression}
}

