% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{liner regression like what lm() does}
\description{
liner regression like what lm() does
}
\section{Fields}{

\describe{
\item{\code{regco:}}{regression coefficient (beta with a hat)}

\item{\code{fitted_value:}}{fitted value (y with a hat)}

\item{\code{residuals}}{: residuals (e with a hat)}

\item{\code{deg_of_fredom}}{: degree of fredom}

\item{\code{resi_var:}}{variance of residuals}

\item{\code{regco_var:}}{variance of regression coefficient}

\item{\code{t_regco:}}{t value of regression coefficient}

\item{\code{object:}}{input formula}

\item{\code{data}}{: input dataframe(the data reference of formula)}

\item{\code{dataname:}}{name of dataframe}

\item{\code{fname}}{: string of formula(without bracket)}

\item{\code{mat1:}}{created by model.matrix()}

\item{\code{y:}}{true value in dataframe not estimated value}
}}


