% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{calculate the statistics using ordinary least squares}
\arguments{
\item{formula}{.}

\item{data}{A dataframe.}
}
\value{
the statistics using ordinary least squares.
}
\description{
calculate the statistics using ordinary least squares
}

\examples{
linreg(formula, data)
}
