% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear Regression}
\arguments{
\item{formula, }{The formula}

\item{data, }{The data, iris}
}
\description{
Run Linear Regression
}
\section{Fields}{

\describe{
\item{\code{formula,}}{formula}

\item{\code{data,}}{data}

\item{\code{x,}}{x}

\item{\code{y,}}{y}

\item{\code{regressions_coefficients,}}{regressions_coefficients}

\item{\code{fitted_values,}}{fitted_values}

\item{\code{the_residuals,}}{the_residuals}

\item{\code{n,}}{n}

\item{\code{p,}}{p}

\item{\code{the_degrees_of_freedom,}}{the_degrees_of_freedom}

\item{\code{the_residual_variance,}}{the_residual_variance}

\item{\code{the_variance_of_the_regression_coefficients,}}{the_variance_of_the_regression_coefficients}

\item{\code{standard_error,}}{standard_error}

\item{\code{sigma1,}}{sigma1}

\item{\code{t_values,}}{t_values}

\item{\code{p_values,}}{p_values}

\item{\code{formula_name,}}{formula_name}

\item{\code{data_name,}}{data_name}
}}


