% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4package.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{linreg}
\arguments{
\item{formula}{a formula}

\item{data}{a data.frame}
}
\value{
attributes and methods of a new RC class
}
\description{
linreg() is a RC class used for create a linear regression
}

