% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Defining the linreg class}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula, e.g. Petal.Length ~ Species.}

\item{data}{A data frame containing the variables in the formula.}
}
\value{
An object of class "linreg" (a list with the coefficients, fitted
values, residuals, degrees of freedom, variances, t-values and p-values).
}
\description{
Defining the linreg class
}
\references{
\url{https://en.wikipedia.org/wiki/Ordinary_least_squares}
}
