% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{The least-square method}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{an object of class "formula" or a character vector on the form y ~ x 
The dependent (y) and independent variable (x)}

\item{data}{a data set from which the used variables are taken (data rejm is prefered)}
}
\value{
an list of class 'linreg' containing the following data:
}
\description{
Computes a linear model using the least-square estimation. 
The function also computes serveral diagnostic measures such as p-value, variance and fitted values.
}
\details{
The 'linreg' function uses ordinary linear algebra to compute the most 
important measures for a linear regression. 
These measures are stored in a list of class 'linreg'.
}
\examples{
## Make a simple linear regression of how the Petal.Length responds to the Petal.Width in the dataset iris.
   linreg(Petal.Length ~ Species, iris)

}
\references{
http://en.wikipedia.org/wiki/Least_squares
http://en.wikipedia.org/wiki/Linear_regression
}

