% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{linreg}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A list representing the adjacency list of the graph.}

\item{data}{The starting node for the algorithm.}
}
\value{
A list containing the shortest paths from the source to all other nodes.
}
\description{
This function implements different linjear regression algorithms.
}
\examples{
linreg(1, 2)

}
\references{
Wikipedia: \href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{Dijkstra's algorithm}
}
