% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.r
\name{linreg}
\alias{linreg}
\title{Linear Regression using QR factorization}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame}
}
\value{
linreg class object
}
\description{
This funcction calculates the parameters for a linear regression for a given
set of data and formula. QR factorization is used to get the final result.
The output is an object of class "linreg"
}
