% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{linreg}
\source{
http://staff.www.ltu.se/~jove/courses/c0002m/least_squares.pdf
}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula stating the form of the regression}

\item{data}{A dataframe to use for the regression}
}
\value{
An object of class linreg containing regression coefficients, fitted values, residuals, degree-of-freedom, residual variance, variance of the regression coefficient, t-values and p-values
}
\description{
A function that performs a linear regression based on QR decomposition
}
