% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linreg.R
\docType{class}
\name{linreg}
\alias{linreg}
\title{Linear Regression Class}
\arguments{
\item{formula}{A formula object}

\item{data}{A dataset}
}
\value{
Returns a linear model built based on \code{formula} and \code{data}
}
\description{
This class is implemented to handle special output
    functions such as print(), plot(), resid(), pred(), coef()
    and summary(). Please refer the methods section of documentation
    for information on each method
}
\details{
Advanced Programming in R - Lab 4
Linear Regression using ordinary least squares
}
\section{Methods}{

\describe{
\item{\code{coef()}}{Return the regression coefficients as a named vector}

\item{\code{is_outlier(x, limit)}}{Return TRUE if data point x is an outlier based on the specified limit}

\item{\code{plot()}}{Plot two graphs - Residuals vs Fitted graph and Scale-Location}

\item{\code{pred()}}{Return the predicted values}

\item{\code{print()}}{Print out the regression coefficients and coefficient names}

\item{\code{resid()}}{Return the vector of residuals}

\item{\code{summary()}}{Print out the regression coefficients with their standard error, t-values and p-values,
along with the estimate of residual standard error and degrees of freedom in the model}
}}

\examples{
\dontrun{
obj = linreg(formula = Petal.Length ~ Species, data = iris)
obj$print()
obj$plot()
}
}
