% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg}
\alias{linreg}
\title{Linear Regression Class}
\usage{
linreg_f(formula, data)
}
\arguments{
\item{formula}{formula (y ~ x)}

\item{data}{dataframe}
}
\value{
Numeric vector representing the shortest distance form
        init_node to all the other nodes in graph
}
\description{
This function finds the shortest distance from a source
             node to all the other nodes in a graph
}
\details{
Advanced Programming in R - Lab 4
Linear Regression using ordinary least squares
}

\examples{
linreg_f(formula = Petal.Length ~ Species, data = iris)
}
