% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{linreg}
\alias{linreg}
\title{Linear Regression Function.}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{The formula for the linear regression with dependent and independent variables.}

\item{data}{The data includes a data set for the linear regression.}
}
\value{
The function returns an object with the class linreg.
}
\description{
The function takes two arguments formula and data. It returns an object with class linreg as an S3 class.
The "y_value_name" extracts the name of the dependent variable.
The "y" has the values of the dependent variable.
The "x" is a matrix which contains independent variables.
The linear regression model is assigned to "model_lr".
The calculation of the p-values of each regression coefficient is done and assigned to "p_values".
Coeffients, call, model and p_values are stored in a list which is called "result".
The function resid returns the vector of residual values of the regression model.
The function pred returns the predicted values y^.
The coef function returns the coefficients as a named vector.
The summary function prints the coefficients with their standard error , t-value,  σ^ and the degrees of freedom in the model.
The class assigns the linreg to the object result.
Creating two graphs , first with Residuals in y axis and Fitted values in the x axis and the second with square root of Standardized residuals in y axis and Fitted values in the x axis, using the function ggplot.
}
