% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression}
\source{
\url{https://en.wikipedia.org/wiki/Linear_regression}
}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{an object of class 'formula'. Describes the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}
}
\value{
The fitted model of class \code{"LinearRegressionModel"}.
}
\description{
'linreg' is used to fit linear regression models.
}
\examples{
linreg(formula = Petal.Width ~ Petal.Length, data=iris)
linreg(Petal.Width ~ Petal.Length + Sepal.Width + Sepal.Length, data=iris)
linreg(Petal.Width ~ Species, iris)

}

