% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Title}
\usage{
linreg(formula, data, method = c("ols", "qr"))
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted}

\item{data}{A data frame containing the variables in the model}

\item{method}{method of OLS or QR}
}
\value{
different methods of function
}
\description{
Title
}
