% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_qr.R
\name{linreg_qr}
\alias{linreg_qr}
\title{Title}
\usage{
linreg_qr(formula, data)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted}

\item{data}{A data frame containing the variables in the model}
}
\value{
statistics
}
\description{
Title
}
