% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linregmod.R
\docType{class}
\name{linregmod}
\alias{.linreg}
\alias{linregmod}
\title{Linear Regression Model}
\source{
\url{https://stat.ethz.ch/R-manual/R-devel/library/methods/html/refClass.html}
}
\arguments{
\item{param}{description}
}
\value{

}
\description{
description here
}
\examples{
examples here

}

