% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot diagnostic plots for a linreg object}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{linreg}.}

\item{...}{Additional arguments.}
}
\description{
Creates diagnostic plots for an object of class \code{linreg}.
The function produces a Residuals vs Fitted plot and a
Scale-Location plot using \code{ggplot2}.
}
