% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plots regarding the residuals}
\usage{
\method{plot}{linreg}(l)
}
\description{
The plot() makes two plots about the relationship between the fitted values and the residuals.
}

