% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.r
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot of linreg object}
\usage{
\method{plot}{linreg}(x)
}
\arguments{
\item{x}{linreg object}
}
\value{
two saved figures
}
\description{
Saves two plot figures
One is the Residuals for each predicted value
The second is the standardized resiuduals for each prediction.
}
