% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot a linreg model}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{extra arguments.}
}
\value{
plot
}
\description{
Plot a linreg model
}
