% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1.R
\name{plot.linreg}
\alias{plot.linreg}
\title{linreg}
\usage{
\method{plot}{linreg}(x, ..)
}
\arguments{
\item{with}{ggplot2 packages}
}
\value{
plot showing residuals vs fitted values
}
\description{
Implementing methods in class
}
