% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{plot S3 method}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class linreg}

\item{...}{other arguments}
}
\value{
List containing the two plots.
}
\description{
S3 method for objects of class linreg providing two plots of the residuals and the standardized residuals among the fitted values
}
