% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Title}
\usage{
\method{plot}{linreg}(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Additional arguments}
}
\value{
p1+p2
}
\description{
Title
}
