% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plots to evaluate the model}
\usage{
plot(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Additional arguments}
}
\value{
The result of the method applied to the class of the input object.
}
\description{
Plots to evaluate the model
}
