% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Extract predicted values from a linreg model}
\usage{
\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector containing the fitted values, with class
  \code{pred}.
}
\description{
Extracts the fitted values from a \code{linreg} object.
}
