% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Predicted values of a linreg object}
\usage{
\method{pred}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class "linreg".}

\item{...}{Not used.}
}
\value{
A vector of predicted values.
}
\description{
Predicted values of a linreg object
}
