% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.linreg.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Available methods in 'linreg'}
\usage{
\method{pred}{linreg}(l, newdata = NA)
}
\description{
The methods available for the linreg-package are described here.
}

