% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.linreg.R
\name{pred.linreg}
\alias{pred.linreg}
\title{The fitted values from the linear regression 'linreg'}
\usage{
\method{pred}{linreg}(l, newdata = NA)
}
\description{
The pred()-statement returns a vector including all fitted values from the linear regression.
}

