% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.r
\name{pred.linreg}
\alias{pred.linreg}
\title{Predictions of linreg object}
\usage{
\method{pred}{linreg}(x)
}
\arguments{
\item{x}{linreg object}
}
\value{
float list
}
\description{
Returns all the predictions of the linreg object
}
