% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred.linreg}
\alias{pred.linreg}
\title{fitted values}
\usage{
\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg object.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector of fitted values.
}
\description{
fitted values
}
