% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred.linreg}
\alias{pred.linreg}
\title{pred S3 method}
\usage{
\method{pred}{linreg}(object)
}
\arguments{
\item{linreg_obj}{An object of class linreg}
}
\value{
vector containing fitted values
}
\description{
S3 method for objects of class linreg providing the fitted values of the linear regression
}
