% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.linreg.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Title}
\usage{
\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Additional arguments}
}
\value{
object$The_fitted_values
}
\description{
Title
}
