% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred}
\alias{pred}
\title{Extract predicted values}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Additional arguments.}
}
\value{
The predicted values.
}
\description{
Generic function for extracting predicted values from a model.
}
