% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pred}
\alias{pred}
\alias{pred.linreg}
\title{Predicted values}
\usage{
pred(x, ...)

\method{pred}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class "linreg".}

\item{...}{Further arguments passed to methods (not used).}
}
\value{
A vector of predicted values.
}
\description{
Returns the predicted values \eqn{\hat{y}} of a fitted model.
}
