% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{The fitted values from the linear regression 'linreg'}
\usage{
pred(l, newdata = NA)
}
\description{
The pred()-statement returns a vector including all fitted values from the linear regression.
}

