% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinReg.R
\name{pred}
\alias{pred}
\alias{pred.linreg}
\title{Predicted/fitted values for a linreg model}
\usage{
pred(object, ...)

\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Ignored.}
}
\value{
A numeric vector of fitted values.
}
\description{
Returns the fitted values from a `linreg` model.
}
