% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred}
\alias{pred}
\title{values}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{extra  arguments.}
}
\value{
A numeric vector of predicted values.
}
\description{
values
}
