% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_pred.R
\name{pred}
\alias{pred}
\title{Generic pred() function for S3}
\usage{
pred(object)
}
\arguments{
\item{object}{some object}
}
\value{
nothing
}
\description{
Provides generic function pred() for S3 objects
}
