% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ridgereg.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Fitted values}
\usage{
\method{predict}{ridgereg}(object, newdata, ...)
}
\arguments{
\item{object}{A model object}

\item{newdata}{newdata}

\item{...}{Additional arguments}
}
\value{
object$The_fitted_values
}
\description{
Fitted values
}
