% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{Print a linreg model}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{linreg}.}

\item{...}{Additional arguments.}
}
\value{
The \code{linreg} object
}
\description{
Prints the function call and estimated regression
coefficients from a \code{linreg} object.
}
