% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.linreg}
\alias{print.linreg}
\title{print function}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class "linreg".}

\item{...}{Not used.}
}
\value{
The object x, invisibly.
}
\description{
print function
}
