% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{The coefficients from the linear regression 'linreg'.}
\usage{
\method{print}{linreg}(l)
}
\description{
The print() returns the names and the regression coefficients.
}

