% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.r
\name{print.linreg}
\alias{print.linreg}
\title{Print linreg class object}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{other options}
}
\description{
Print the coefficients and the call function of the linear regression
}
