% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{Print a linreg model}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg object.}

\item{...}{extra arguments.}
}
\value{
A vector of regression coefficients
}
\description{
Print a linreg model
}
