% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linreg}
\alias{print.linreg}
\title{print S3 method}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class linreg}

\item{...}{other arguments}
}
\value{
nothing. Prints tha call and the coefficients of the linreg object.
}
\description{
S3 method for objects of class linreg printing the call and the coefficients of the linear model
}
