% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{print linreg}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg class object}

\item{...}{other parameters}
}
\description{
Print the linreg class
}
