% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{Simple overview of the fitted model, showing important coefficients and statistics}
\usage{
\method{print}{linreg}(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Additional arguments}
}
\value{
print the formula, data and coefficients
}
\description{
Simple overview of the fitted model, showing important coefficients and statistics
}
