% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ridgereg.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Simple overview of the fitted model, showing important coefficients and statistics}
\usage{
\method{print}{ridgereg}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{Additional arguments}
}
\value{
print the formula, data and coefficients
}
\description{
Simple overview of the fitted model, showing important coefficients and statistics
}
