% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Extract residuals}
\usage{
\method{resid}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector containing the residuals.
}
\description{
Extracts the residuals from a \code{linreg} object.
}
