% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.linreg.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Extracts the residuals from the fitted model}
\usage{
\method{resid}{linreg}(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Additional arguments}
}
\value{
object$The_residuals
}
\description{
Extracts the residuals from the fitted model
}
