% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{resid() should return the vector of residuals eˆ}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class "linreg".}

\item{...}{Not used.}
}
\value{
A vector of residuals.
}
\description{
resid() should return the vector of residuals eˆ
}
