% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.linreg.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Available methods in 'linreg'}
\usage{
\method{residuals}{linreg}(l)
}
\description{
The methods available for the linreg-package are described here.
}

