% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.linreg.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{The residuals from the linear regression 'linreg'}
\usage{
\method{residuals}{linreg}(l)
}
\description{
The resid() returns all the residuals from the model in a vector.
}

