% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.r
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Residuals of linreg object}
\usage{
\method{residuals}{linreg}(x)
}
\arguments{
\item{x}{linreg object}
}
\value{
float list
}
\description{
Returns all the residuals of the linreg object
}
