% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Ridge regression}
\usage{
ridgereg(formula, lambda = 0, data)
}
\arguments{
\item{formula}{an object of class "formula" or a character vector on the form y ~ x 
The dependent (y) and independent variable (x)}

\item{data}{a data set from which the used variables are taken}
}
\value{
an list of class 'ridgereg' containing beta coefficients and fitted values.
}
\description{
Computes a ridge model using the Ridge -square method
The function also fitted values.
}
\details{
These measures are stored in a list of class 'ridgereg'.
}
\examples{
## Make a simple linear regression of how the Petal.Length responds to the Petal.Width in the dataset iris.
   ridgereg(Petal.Length ~ Species, lambda = 5, iris)

}
\references{

}

