% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Linear regression with ridge}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted}

\item{data}{A data frame containing the variables in the model}

\item{lambda}{Regularization parameter}
}
\value{
statistics
}
\description{
Linear regression with ridge
}
