% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summarize a linreg model}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Additional arguments.}
}
\value{
The \code{linreg} object, invisibly.
}
\description{
Prints a summary of a \code{linreg} object, including the
estimated regression coefficients, standard errors,
t-statistics, p-values, residual standard error, and residual
degrees of freedom.
}
