% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.linreg}
\alias{summary.linreg}
\title{summary}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class "linreg".}

\item{...}{Not used.}
}
\value{
The object, invisibly.
}
\description{
summary
}
\references{
\url{https://en.wikipedia.org/wiki/T-statistic}
}
