% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.linreg.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary of the linear regression}
\usage{
\method{summary}{linreg}(l)
}
\description{
summary provides the most important measures and values from the linear regression.
}

