% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary of a linear model}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{linreg object.}

\item{...}{extra arguments.}
}
\value{
Prints the model summary.
}
\description{
Summary of a linear model
}
