% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.linreg}
\alias{summary.linreg}
\title{summary S3 method}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{...}{other arguments}
}
\value{
nothing. Prints a summary of the linreg object.
}
\description{
S3 method for objects of class linreg providing a summary of the object
}
