% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{summary.linreg}
\alias{summary.linreg}
\title{summary linreg}
\usage{
\method{summary}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg class object}

\item{...}{other parameters}
}
\description{
print the summary of the linreg class object
}
