% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.linreg.R
\name{summary.linreg}
\alias{summary.linreg}
\title{a detailed report of the model’s coefficients and other statistics.}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Additional arguments}
}
\value{
print the sum
}
\description{
a detailed report of the model’s coefficients and other statistics.
}
