% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize aiport delays}
\usage{
visualize_airport_delays()
}
\description{
Visualize all airport delays in the nycflights13 package with a mapogram.
The package takes no arguments.
}

